address attrocious styling of previous code
